+++ cmd/enumeration/run.sh
