+++ docker-compose-test.yml
@@ -31,9 +31,18 @@ services:
       MYSQL_DATABASE: scanner
       MYSQL_ROOT_PASSWORD: root
     ports:
-      - '3307:3306'
-    volumes:
-      - db:/var/lib/mysql
-volumes:
-  db:
-    driver: local
+      - '3317:3306'
+    healthcheck:
+      test: ['CMD', 'mysqladmin', 'ping', '-h', 'localhost', '--user=root', '--password=root']
+      interval: 3s
+      timeout: 3s
+      retries: 5
+  waiter: # a noop container to wait for the services to be up before we run tests
+    image: mysql:8.0
+    command: bash -c "exit 0"
+    restart: "no"
+    depends_on:
+      broker:
+        condition: "service_started" 
+      db:
+        condition: "service_healthy" # avoids "EOF" in tests
