--- enumeration/dto/ingest.go
@@ -1,6 +1,0 @@
-package dto
-
-type IngestDto struct {
-	Domain string
-	Id     uint32
-}
