--- enumeration/dto/root_domain.go
@@ -1,9 +1,0 @@
-package dto
-
-type RootDomainDto struct {
-	Id     uint32
-	Root   string
-	Domain string
-	Status string
-	Owner  string
-}
