--- enumeration/go.mod
@@ -1,50 +1,0 @@
-module enumeration
-
-go 1.17
-
-require (
-	github.com/confluentinc/confluent-kafka-go v1.9.2
-	github.com/projectdiscovery/subfinder/v2 v2.5.3
-)
-
-require (
-	github.com/andres-erbsen/clock v0.0.0-20160526145045-9e14626cd129 // indirect
-	github.com/asaskevich/govalidator v0.0.0-20210307081110-f21760c49a8d // indirect
-	github.com/aymerick/douceur v0.2.0 // indirect
-	github.com/cnf/structhash v0.0.0-20201127153200-e1b16c1ebc08 // indirect
-	github.com/corpix/uarand v0.2.0 // indirect
-	github.com/gorilla/css v1.0.0 // indirect
-	github.com/hako/durafmt v0.0.0-20210316092057-3a2c319c1acd // indirect
-	github.com/json-iterator/go v1.1.12 // indirect
-	github.com/lib/pq v1.10.6 // indirect
-	github.com/logrusorgru/aurora v2.0.3+incompatible // indirect
-	github.com/microcosm-cc/bluemonday v1.0.18 // indirect
-	github.com/miekg/dns v1.1.46 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v1.0.2 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/projectdiscovery/blackrock v0.0.0-20210903102120-5a9d2412d21d // indirect
-	github.com/projectdiscovery/chaos-client v0.2.0 // indirect
-	github.com/projectdiscovery/dnsx v1.1.0 // indirect
-	github.com/projectdiscovery/fileutil v0.0.0-20220625080234-94e882ca56cb // indirect
-	github.com/projectdiscovery/goflags v0.0.9 // indirect
-	github.com/projectdiscovery/gologger v1.1.4 // indirect
-	github.com/projectdiscovery/httputil v0.0.0-20210906072657-f3a099cb20bc // indirect
-	github.com/projectdiscovery/iputil v0.0.0-20210804143329-3a30fcde43f3 // indirect
-	github.com/projectdiscovery/mapcidr v0.0.8 // indirect
-	github.com/projectdiscovery/retryabledns v1.0.13 // indirect
-	github.com/projectdiscovery/retryablehttp-go v1.0.2 // indirect
-	github.com/projectdiscovery/sliceutil v0.0.0-20220426000009-1d2b7c02f65c // indirect
-	github.com/projectdiscovery/stringsutil v0.0.0-20220612082425-0037ce9f89f3 // indirect
-	github.com/rs/xid v1.4.0 // indirect
-	github.com/saintfish/chardet v0.0.0-20120816061221-3af4cd4741ca // indirect
-	github.com/tomnomnom/linkheader v0.0.0-20180905144013-02ca5825eb80 // indirect
-	go.uber.org/ratelimit v0.2.0 // indirect
-	golang.org/x/mod v0.5.1 // indirect
-	golang.org/x/net v0.0.0-20220624214902-1bab6f366d9e // indirect
-	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
-	golang.org/x/text v0.3.7 // indirect
-	golang.org/x/tools v0.1.9 // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
