--- enumeration/handlers/server.go
@@ -1,8 +1,0 @@
-package handlers
-
-import "github.com/confluentinc/confluent-kafka-go/kafka"
-
-type Server struct {
-	Producer *kafka.Producer
-	Consumer *kafka.Consumer
-}
