--- orchestrator/dto/kafka_message.go
@@ -1,6 +1,0 @@
-package dto
-
-type KafkaMessage struct {
-	Type    string
-	Payload map[string]interface{}
-}
