--- orchestrator/dto/root_domain.go
@@ -1,8 +1,0 @@
-package dto
-
-type RootDomainDto struct {
-	Id     uint32
-	Root   string
-	Status string
-	Owner  string
-}
