--- orchestrator/dto/subdomain.go
@@ -1,7 +1,0 @@
-package dto
-
-type SubdomainDto struct {
-	Id     uint32
-	Root   string
-	Source string
-}
