--- orchestrator/go.mod
@@ -1,8 +1,0 @@
-module orchestrator
-
-go 1.17
-
-require (
-	github.com/confluentinc/confluent-kafka-go v1.9.2
-	github.com/go-sql-driver/mysql v1.6.0
-)
