--- orchestrator/handlers/server.go
@@ -1,16 +1,0 @@
-package handlers
-
-import (
-	"database/sql"
-	"github.com/confluentinc/confluent-kafka-go/kafka"
-)
-
-type Server struct {
-	Producer  *kafka.Producer
-	Consumer  *kafka.Consumer
-	SqlClient SqlClient
-}
-
-type SqlClient struct {
-	DB *sql.DB
-}
